v1.0
Complete and efficient calculation of a single rounds data complete with visualization.
