+++ Legacy Files/Utils/CollisionDetection.py
