+++ Legacy Files/Utils/PositionGenerator.py
