--- NoSimulator.py
@@ -1,77 +1,0 @@
-import numpy as np
-import time
-import matplotlib.pyplot as plt
-from matplotlib.animation import FuncAnimation
-
-from Simulation import NextMove, FoodCollision, ConstructOrganismArray, ConstructFoodArray
-
-
-def RunRound(organismArray, foodArray, frameRate):
-    # * Simulation Conditions
-
-    roundLength = 60 * frameRate + 1
-
-    foodArrayDisplay = foodArray.copy()
-
-    desired_interval = 2.5  # Desired interval in second
-    directionChangeProbability = 1 - \
-        np.exp(-1 / (frameRate * desired_interval))
-
-    for roundNumber in range(1, roundLength):
-        start_time = time.time()
-        organismArray = NextMove(
-            organismArray, roundNumber, boundaryRadius, directionChangeProbability)
-        elapsed_time = time.time() - start_time
-        print(
-            f"Frame-{roundNumber + 1} movement took {elapsed_time:.6f} seconds to simulate.")
-        organismArray, foodArray = FoodCollision(
-            organismArray, foodArray, roundNumber)
-        elapsed_time = time.time() - start_time
-        print(
-            f"Frame-{roundNumber + 1} collision took {elapsed_time:.6f} seconds to simulate.")
-        time.sleep(1)
-
-    percentConsumed = (
-        np.sum(organismArray[5, :, -1])/foodArrayDisplay.shape[1]) * 100
-    print(np.sum(organismArray[5, :, -1]))
-
-    print(f"Percentage of Food Consumed: {percentConsumed}%")
-    return organismArray, foodArrayDisplay
-
-
-def visualize(tuple, frameRate):
-    simulationMatrix, foodArray = tuple
-    time_steps = simulationMatrix.shape[2]
-    fig, ax = plt.subplots()
-    sc = ax.scatter([], [], c='b', marker='D')
-    ax.set_title('Animated Scatter Plot')
-    ax.set_xlabel('X Position')
-    ax.set_ylabel('Y Position')
-    ax.set_xlim(-55, 55)  # Adjust limits based on your data
-    ax.set_ylim(-55, 55)  # Adjust limits based on your data
-
-    ax2 = ax.twinx()
-    ax2.scatter(foodArray[0, :], foodArray[1, :],
-                c='r', marker='x', label='Tree')
-    ax2.legend()
-
-    def animate(frame):
-        # Access the 3D matrix for the current time step
-        current_positions = simulationMatrix[:, :, frame]
-        # Plot the scatter plot with current positions
-        sc.set_offsets(current_positions[:2, :].T)
-        ax.set_title(
-            f'Food Consumed: {np.sum(current_positions[5,:])}  - F{frame}')
-
-    animation = FuncAnimation(
-        fig, animate, frames=time_steps, interval=1000/frameRate)
-
-    plt.show()
-
-
-if __name__ == "__main__":
-    frameRate = 60
-    boundaryRadius = 50
-    organismArray = ConstructOrganismArray(12000, boundaryRadius, frameRate)
-    foodArray = ConstructFoodArray(200, boundaryRadius)
-    visualize(RunRound(organismArray, foodArray, frameRate), frameRate)
